+++ grow/apps.py
@@ -4,6 +4,3 @@ from django.apps import AppConfig
 class GrowConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "grow"
-
-    def ready(self):
-        import grow.model.user.signals