+++ grow/model/play/Play.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
@@ -16,6 +17,10 @@ class Play(PrivateUniqueResource):
     Uses Django's generic foreign key mechanism to reference the played content,
     which can be either a Track (e.g. YoutubeTrack) or a Playlist.
     """
+
+    user = models.ForeignKey(
+        settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name="%(class)ss", null=True, blank=True
+    )
 
     content_type = AppForeignKey(ContentType, on_delete=models.CASCADE)
     content_uuid = models.UUIDField(db_column="object_pk")